Moving update_column_width into alignment_utils file.
