+++ vsg/rules/align_tokens_in_region_between_tokens_unless_between_tokens.py
@@ -4,9 +4,10 @@ from vsg import parser
 from vsg import token
 from vsg import violation
 
-from vsg.vhdlFile import utils
 from vsg.rule_group import alignment
+from vsg.rules import alignment_utils
 from vsg.rules import utils as rule_utils
+from vsg.vhdlFile import utils
 
 
 class align_tokens_in_region_between_tokens_unless_between_tokens(alignment.Rule):
@@ -91,7 +92,7 @@ class align_tokens_in_region_between_tokens_unless_between_tokens(alignment.Rule
                                dAnalysis[iLine]['left_column'] = iColumn
                            break
 
-                   iColumn += update_column_width(self, oToken)
+                   iColumn += alignment_utils.update_column_width(self, oToken)
 
                if isinstance(oToken, parser.carriage_return):
                    iLine += 1
@@ -316,13 +317,3 @@ def is_case_keyword(config, iIndex, lTokens):
 
 def get_tokens_of_interest(self, oFile):
     return oFile.get_tokens_bounded_by_unless_between(self.left_token, self.right_token, self.lUnless)
-
-
-def update_column_width(self, oToken):
-    sToken = oToken.get_value()
-    if isinstance(oToken, parser.whitespace):
-        iTabs = sToken.count('\t')
-        iLength = len(sToken) + (iTabs * self.indentSize) - iTabs
-        return iLength
-
-    return len(oToken.get_value())
